GenGIF/requirements.txt
matplotlib
Pillow
argparse
